requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
telethon==1.37.0
cryptg==0.4.0
pydantic-settings==2.7.1
